+++ e2e/multicast_publisher_test.go
@@ -216,17 +216,6 @@ func checkMulticastPublisherPostConnect(t *testing.T, dn *TestDevnet, device *de
 			t.Fail()
 		}
 
-		if !t.Run("only_one_tunnel_allowed", func(t *testing.T) {
-			// Set access pass for the client.
-			_, err = dn.Manager.Exec(t.Context(), []string{"bash", "-c", "doublezero access-pass set --accesspass-type prepaid --client-ip " + client.CYOANetworkIP + " --user-payer " + client.Pubkey})
-			require.NoError(t, err)
-
-			_, err := client.Exec(t.Context(), []string{"bash", "-c", "doublezero connect ibrl --client-ip " + client.CYOANetworkIP})
-			require.Error(t, err, "User with different type already exists. Only one tunnel currently supported")
-		}) {
-			t.Fail()
-		}
-
 		dn.log.Info("--> Multicast publisher post-connect requirements checked")
 	})
 }